miApiJWT/requirements.txt
fastapi
uvicorn[standard]
flask
requests
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart